+++ android/app/src/main/java/com/on/turip/data/turip/repository/DefaultTuripRepository.kt
@@ -3,6 +3,7 @@ package com.on.turip.data.turip.repository
 import com.on.turip.core.result.TuripResult
 import com.on.turip.core.result.mapCatching
 import com.on.turip.data.turip.datasource.TuripRemoteDataSource
+import com.on.turip.data.turip.datasource.TuripSseStreamDataSource
 import com.on.turip.data.turip.dto.PlaceTuripsRequest
 import com.on.turip.data.turip.dto.TuripPlaceOrderRequest
 import com.on.turip.data.turip.toDomain
@@ -18,53 +19,54 @@ import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
 
 class DefaultTuripRepository @Inject constructor(
-    private val turipRemoteDataSource: TuripRemoteDataSource,
+    private val turipRestRemoteDataSource: TuripRemoteDataSource,
+    private val turipSseStreamDataSource: TuripSseStreamDataSource,
 ) : TuripRepository {
     override suspend fun loadTurip(turipId: Long): TuripResult<Turip> =
-        turipRemoteDataSource.getTurip(turipId).mapCatching { it.toDomain() }
+        turipRestRemoteDataSource.getTurip(turipId).mapCatching { it.toDomain() }
 
-    override suspend fun loadTurips(): TuripResult<List<Turip>> = turipRemoteDataSource.getTurips().mapCatching { it.toDomain() }
+    override suspend fun loadTurips(): TuripResult<List<Turip>> = turipRestRemoteDataSource.getTurips().mapCatching { it.toDomain() }
 
     override suspend fun loadTuripMembers(turipId: Long): TuripResult<List<String>> =
-        turipRemoteDataSource.getTuripMembers(turipId).mapCatching { it.toDomain() }
+        turipRestRemoteDataSource.getTuripMembers(turipId).mapCatching { it.toDomain() }
 
     override suspend fun createTurip(name: String): TuripResult<Turip> =
-        turipRemoteDataSource
+        turipRestRemoteDataSource
             .postTurip(name.toPostRequestDto())
             .mapCatching { it.toDomain() }
 
     override suspend fun updateTurip(
         turipId: Long,
         updateName: String,
-    ): TuripResult<Unit> = turipRemoteDataSource.patchTurip(turipId, updateName.toPatchRequestDto())
+    ): TuripResult<Unit> = turipRestRemoteDataSource.patchTurip(turipId, updateName.toPatchRequestDto())
 
-    override suspend fun deleteTurip(turipId: Long): TuripResult<Unit> = turipRemoteDataSource.deleteTurip(turipId)
+    override suspend fun deleteTurip(turipId: Long): TuripResult<Unit> = turipRestRemoteDataSource.deleteTurip(turipId)
 
-    override suspend fun exitTurip(turipId: Long): TuripResult<Unit> = turipRemoteDataSource.exitTurip(turipId)
+    override suspend fun exitTurip(turipId: Long): TuripResult<Unit> = turipRestRemoteDataSource.exitTurip(turipId)
 
     override suspend fun loadTuripsByPlaceId(placeId: Long): TuripResult<List<Turip>> =
-        turipRemoteDataSource
+        turipRestRemoteDataSource
             .getTuripsByPlaceId(placeId)
             .mapCatching { it.toDomain() }
 
     override suspend fun loadTuripPlaces(turipId: Long): TuripResult<List<TuripPlace>> =
-        turipRemoteDataSource.getTuripPlaces(turipId).mapCatching { it.toDomain() }
+        turipRestRemoteDataSource.getTuripPlaces(turipId).mapCatching { it.toDomain() }
 
     override suspend fun createTuripPlace(
         turipId: Long,
         placeId: Long,
-    ): TuripResult<Unit> = turipRemoteDataSource.createTuripPlace(turipId = turipId, placeId = placeId)
+    ): TuripResult<Unit> = turipRestRemoteDataSource.createTuripPlace(turipId = turipId, placeId = placeId)
 
     override suspend fun deleteTuripPlace(
         turipId: Long,
         placeId: Long,
-    ): TuripResult<Unit> = turipRemoteDataSource.deleteTuripPlace(turipId = turipId, placeId = placeId)
+    ): TuripResult<Unit> = turipRestRemoteDataSource.deleteTuripPlace(turipId = turipId, placeId = placeId)
 
     override suspend fun updateTuripPlacesOrder(
         turipId: Long,
         updatedOrder: List<Long>,
     ): TuripResult<Unit> =
-        turipRemoteDataSource
+        turipRestRemoteDataSource
             .patchTuripPlacesOrder(
                 turipId = turipId,
                 turipPlaceOrderRequest = TuripPlaceOrderRequest(turipPlaceIdsOrder = updatedOrder),
@@ -74,18 +76,18 @@ class DefaultTuripRepository @Inject constructor(
         placeId: Long,
         turipIds: List<Long>,
     ): TuripResult<Unit> =
-        turipRemoteDataSource.putPlaceTurips(
+        turipRestRemoteDataSource.putPlaceTurips(
             placeId = placeId,
             placeTuripsRequest = PlaceTuripsRequest(turipIds),
         )
 
     override suspend fun createInvitationToken(turipId: Long): TuripResult<TuripInvitationToken> =
-        turipRemoteDataSource.createInvitationToken(turipId).mapCatching { it.toDomain() }
+        turipRestRemoteDataSource.createInvitationToken(turipId).mapCatching { it.toDomain() }
 
-    override suspend fun joinTurip(turipId: Long): TuripResult<Unit> = turipRemoteDataSource.joinTurip(turipId).mapCatching { Unit }
+    override suspend fun joinTurip(turipId: Long): TuripResult<Unit> = turipRestRemoteDataSource.joinTurip(turipId).mapCatching { Unit }
 
     override suspend fun verifyInvitationToken(token: String): TuripResult<TuripInvitationInformation> =
-        turipRemoteDataSource.getInvitationInformation(token).mapCatching { it.toDomain() }
+        turipRestRemoteDataSource.getInvitationInformation(token).mapCatching { it.toDomain() }
 
-    override fun streamTuripEvents(turipId: Long): Flow<TuripResult<TuripStreamEvent>> = turipRemoteDataSource.streamTuripEvents(turipId)
+    override fun streamTuripEvents(turipId: Long): Flow<TuripResult<TuripStreamEvent>> = turipSseStreamDataSource.streamTuripEvents(turipId)
 }
